🚚 Rename: context 폴더로 AuthSession 파일 이동
